Django_app/Django_app/requirements.txt
asgiref==3.4.0
Django==3.2.4
django-crispy-forms==1.12.0
django-model-utils==4.1.1
pyotp==2.6.0
pytz==2021.1
sqlparse==0.4.1

